cpu_input_boost: Allow configuration of each cluster's max-boost freq
On certain CPUs, the highest frequency could be very inefficient, and
thus frequently boosting to the frequency for max boosts can incur a
high power cost. To mitigate this, allow configuration of the max-boost
frequency for each CPU cluster, so that a more efficient frequency can
be selected for max boosts.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Akshyay Menon <menon_akshay@live.com>